requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
cliff>=1.10.0  # Apache-2.0
jsonschema>=2.0.0,<3.0.0
oslo.utils>=1.4.0                       # Apache-2.0
pbr>=0.6,!=0.7,<1.0
python-keystoneclient>=1.1.0
requests>=2.2.0,!=2.4.0
six>=1.9.0
stevedore>=1.3.0  # Apache-2.0
